small tidy up
