+++ src/formatters/utils.js
@@ -19,46 +19,6 @@ export const extensions = {
 };
 
 /**
- * Partition a network as needed for edge-list and adjacency-matrix formats.
- * Each network contains a reference to the original nodes, with a subset of edges
- * based on the type.
- *
- * @param  {Array} network in NC format
- * @param  {string} format one of `formats`
- * @return {Array} An array of networks, partitioned by edge type. Each network object is decorated
- *                 with an additional `edgeType` prop to facilitate format naming.
- */
-export const partitionByEdgeType = (network, format) => {
-  switch (format) {
-    case 'graphml':
-    case 'ego':
-    case 'attributeList':
-      return [network];
-    case 'edgeList':
-    case 'adjacencyMatrix': {
-      if (!network.edges.length) {
-        return [network];
-      }
-
-      const { nodes } = network;
-      const partitionedEdgeMap = network.edges.reduce((edgeMap, edge) => {
-        edgeMap[edge.type] = edgeMap[edge.type] || []; // eslint-disable-line no-param-reassign
-        edgeMap[edge.type].push(edge);
-        return edgeMap;
-      }, {});
-
-      return Object.entries(partitionedEdgeMap).map(([edgeType, edges]) => ({
-        nodes,
-        edges,
-        edgeType,
-      }));
-    }
-    default:
-      throw new Error('Unexpected format', format);
-  }
-};
-
-/**
  * Provide the appropriate file extension for the export type
  * @param  {string} formatterType one of the `format`s
  * @return {string}
